+++ essencium-backend/src/main/java/de/frachtwerk/essencium/backend/controller/AuthenticationController.java
@@ -30,14 +30,26 @@ import de.frachtwerk.essencium.backend.security.JwtTokenAuthenticationFilter;
 import de.frachtwerk.essencium.backend.security.event.CustomAuthenticationSuccessEvent;
 import de.frachtwerk.essencium.backend.service.JwtTokenService;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.enums.ParameterIn;
+import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.annotation.Nullable;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.validation.constraints.NotNull;
 import java.io.Serializable;
-import java.util.*;
+import java.net.URI;
+import java.net.URISyntaxException;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Objects;
+import java.util.Set;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.http.HttpHeaders;
@@ -59,6 +71,7 @@ import org.springframework.web.server.ResponseStatusException;
     havingValue = "false",
     matchIfMissing = true)
 @Tag(name = "AuthenticationController", description = "Set of endpoints used for authentication")
+@Slf4j
 @RequiredArgsConstructor
 public class AuthenticationController {
   private final AppProperties appProperties;
@@ -75,7 +88,9 @@ public class AuthenticationController {
   }
 
   @PostMapping("/token")
-  @Operation(description = "Log in to request a new JWT token")
+  @Operation(
+      summary = "Login with username and password",
+      description = "Log in to request a new JWT token using username and password")
   public TokenResponse postLogin(
       @RequestBody @Validated LoginRequest login,
       @RequestHeader(value = HttpHeaders.USER_AGENT, required = false) String userAgent,
@@ -115,7 +130,9 @@ public class AuthenticationController {
 
   @PostMapping("/renew")
   @CrossOrigin(origins = "${app.url}", allowCredentials = "true")
-  @Operation(description = "Request a new JWT access token, given a valid refresh token")
+  @Operation(
+      summary = "Renew access token",
+      description = "Request a new JWT access token, given a valid refresh token")
   public TokenResponse postRenew(
       @RequestHeader(value = HttpHeaders.USER_AGENT) String userAgent,
       @CookieValue(value = "refreshToken") String refreshToken,
@@ -144,6 +161,10 @@ public class AuthenticationController {
   }
 
   @GetMapping("/oauth-registrations")
+  @Operation(
+      summary = "Get OAuth2 client registrations",
+      description =
+          "Returns a map of OAuth2 client registrations with their names, URLs, and images")
   public Map<String, Map<String, Object>> getRegistrations() {
     if (!appOAuth2Properties.isEnabled()
         || Objects.isNull(oAuth2ClientRegistrationProperties.getRegistration())) {
@@ -185,12 +206,62 @@ public class AuthenticationController {
     return map;
   }
 
+  @PostMapping("/logout")
+  @Parameter(
+      in = ParameterIn.QUERY,
+      name = "redirectUrl",
+      description = "URL to redirect to after logout",
+      schema = @Schema(type = "string", format = "uri", example = "https://example.com/logout"))
+  @Operation(summary = "Logout the currently logged-in user")
+  public void logout(
+      @RequestHeader(value = HttpHeaders.AUTHORIZATION) @NotNull final String authorizationHeader,
+      @RequestParam(value = "redirectUrl", required = false) @Nullable String redirectUrl,
+      @NotNull HttpServletResponse response) {
+    if (StringUtils.isBlank(redirectUrl)) {
+      redirectUrl = appProperties.getDefaultLogoutRedirectUrl();
+    }
+    // Create URI from redirect URL to validate it as a valid URI
+    URI redirectUri = null;
+    try {
+      redirectUri = new URI(redirectUrl);
+    } catch (URISyntaxException e) {
+      log.error("Invalid redirect URL: {}", redirectUrl, e);
+      jwtTokenService.logout(
+          authorizationHeader, null, oAuth2ClientRegistrationProperties, response);
+      return;
+    }
+
+    if (!isRedirectUrlAllowed(redirectUri)) {
+      log.error("Invalid redirect URL: {}", redirectUrl);
+      jwtTokenService.logout(
+          authorizationHeader, null, oAuth2ClientRegistrationProperties, response);
+      return;
+    }
+
+    jwtTokenService.logout(
+        authorizationHeader, redirectUri, oAuth2ClientRegistrationProperties, response);
+  }
+
   @RequestMapping(value = "/**", method = RequestMethod.OPTIONS)
+  @Operation(
+      summary = "Collection options",
+      description = "Returns the allowed HTTP methods for the authentication endpoints.")
   public final ResponseEntity<?> collectionOptions() {
     return ResponseEntity.ok().allow(getAllowedMethods().toArray(new HttpMethod[0])).build();
   }
 
   protected Set<HttpMethod> getAllowedMethods() {
     return Set.of(HttpMethod.HEAD, HttpMethod.POST, HttpMethod.OPTIONS);
+  }
+
+  private boolean isRedirectUrlAllowed(URI redirectUrl) {
+    return appProperties.getAllowedLogoutRedirectUrls().stream()
+        .anyMatch(
+            allowedUrl -> {
+              String redirectUrlString = redirectUrl.toString();
+              return allowedUrl.contains("*")
+                  ? redirectUrlString.matches(allowedUrl.replace("*", ".*"))
+                  : redirectUrlString.equals(allowedUrl);
+            });
   }
 }